CMake: produce dynamic library and associated python module (Thanks to neomilium)
Previously, CMake produces a python module with whole libarea sources in it; with this patch it now produces a dynamic library which can be used without python and a python wrapper that is linked to the library.
Additionally, CMake now installs headers, allowing developers to include them in their own program which have to be linked dynamically with the library (.so).

Author: Romuald Conty <neomilium@gmail.com>